requirements.txt
wincertstore==0.2
django
djangorestframework